+++ apps/client/src/shared/code-highlighter/utils/parseHighlightHtml.ts
